backend1/requirements.txt
flask==3.0.2
google-generativeai==0.7.2
requests==2.31.0

